annotator/scripts/requirements.txt
Pillow>=10.0
requests>=2.28
numpy>=1.24
